+++ packages/coverde_cli/lib/src/utils/path.dart
@@ -1,4 +1,0 @@
-import 'package:path/path.dart' as p;
-
-/// Standard path context for path manipulation.
-final path = p.Context(style: p.Style.url);